+++ packages/core/src/__tests__/ceramic-anchor.test.ts
@@ -1,44 +1,19 @@
 import { jest } from '@jest/globals'
 import { Ceramic } from '../ceramic.js'
-import { Ed25519Provider } from 'key-did-provider-ed25519'
 import { AnchorStatus, IpfsApi, TestUtils } from '@ceramicnetwork/common'
-import tmp from 'tmp-promise'
-import * as u8a from 'uint8arrays'
 import { createIPFS, swarmConnect } from '@ceramicnetwork/ipfs-daemon'
 import { TileDocument } from '@ceramicnetwork/stream-tile'
 import { InMemoryAnchorService } from '../anchor/memory/in-memory-anchor-service.js'
 import { anchorUpdate } from '../state-management/__tests__/anchor-update.js'
-import * as ThreeIdResolver from '@ceramicnetwork/3id-did-resolver'
-import * as KeyDidResolver from 'key-did-resolver'
-import { Resolver } from 'did-resolver'
-import { DID } from 'dids'
+import { createCeramic as vanillaCreateCeramic } from './create-ceramic.js'
 
-const seed = u8a.fromString(
-  '6e34b2e1a9624113d81ece8a8a22e6e97f0e145c25c1d4d2d0e62753b4060c83',
-  'base16'
-)
-
-const makeDID = function (seed: Uint8Array, ceramic: Ceramic): DID {
-  const provider = new Ed25519Provider(seed)
-
-  const keyDidResolver = KeyDidResolver.getResolver()
-  const threeIdResolver = ThreeIdResolver.getResolver(ceramic)
-  const resolver = new Resolver({
-    ...threeIdResolver,
-    ...keyDidResolver,
-  })
-  return new DID({ provider, resolver })
-}
+const SEED = '6e34b2e1a9624113d81ece8a8a22e6e97f0e145c25c1d4d2d0e62753b4060c83'
 
 const createCeramic = async (ipfs: IpfsApi, anchorManual: boolean): Promise<Ceramic> => {
-  const ceramic = await Ceramic.create(ipfs, {
-    stateStoreDirectory: await tmp.tmpName(),
+  return vanillaCreateCeramic(ipfs, {
     anchorOnRequest: !anchorManual,
-    pubsubTopic: '/ceramic/inmemory/test', // necessary so Ceramic instances can talk to each other
+    seed: SEED,
   })
-  await ceramic.setDID(makeDID(seed, ceramic))
-
-  return ceramic
 }
 
 describe('Ceramic anchoring', () => {