+++ packages/stream-model-instance-handler/src/__tests__/model-instance-document-handler.test.ts
@@ -5,11 +5,11 @@ import * as dagCBOR from '@ipld/dag-cbor'
 import type { DID } from 'dids'
 import { wrapDocument } from '@ceramicnetwork/3id-did-resolver'
 import * as KeyDidResolver from 'key-did-resolver'
-import { TileDocumentHandler } from '../tile-document-handler.js'
+import { ModelInstanceDocumentHandler } from '../model-instance-document-handler.js'
 import * as uint8arrays from 'uint8arrays'
 import * as sha256 from '@stablelib/sha256'
 import cloneDeep from 'lodash.clonedeep'
-import { TileDocument } from '@ceramicnetwork/stream-tile'
+import { ModelInstanceDocument } from '@ceramicnetwork/stream-model-instance'
 import {
   CeramicApi,
   CommitType,
@@ -234,9 +234,9 @@ const rotateKey = (did: DID, rotateDate: string) => {
   did.createJWS = async () => jwsForVersion1
 }
 
-describe('TileDocumentHandler', () => {
+describe('ModelInstanceDocumentHandler', () => {
   let did: DID
-  let tileDocumentHandler: TileDocumentHandler
+  let modelInstanceDocumentHandler: ModelInstanceDocumentHandler
   let context: Context
   let signerUsingNewKey: CeramicSigner
   let signerUsingOldKey: CeramicSigner
@@ -294,17 +294,17 @@ describe('TileDocumentHandler', () => {
   })
 
   beforeEach(() => {
-    tileDocumentHandler = new TileDocumentHandler()
+    modelInstanceDocumentHandler = new ModelInstanceDocumentHandler()
 
     setDidToNotRotatedState(did)
   })
 
   it('is constructed correctly', async () => {
-    expect(tileDocumentHandler.name).toEqual('tile')
+    expect(modelInstanceDocumentHandler.name).toEqual('MID')
   })
 
   it('makes genesis commits correctly', async () => {
-    const commit = await TileDocument.makeGenesis(context.api, COMMITS.genesis.data)
+    const commit = await ModelInstanceDocument.makeGenesis(context.api, COMMITS.genesis.data)
     expect(commit).toBeDefined()
 
     const { jws, linkedBlock } = commit as SignedCommitContainer
@@ -329,19 +329,13 @@ describe('TileDocumentHandler', () => {
     expect(serialized).toEqual(signed)
   })
 
-  it('throws error for deterministic genesis commit with data', async () => {
-    await expect(
-      TileDocument.makeGenesis(context.api, COMMITS.genesis.data, { deterministic: true })
-    ).rejects.toThrow(/Initial content must be null/)
-  })
-
-  it('Does not sign commit if no content', async () => {
-    const commit = (await TileDocument.makeGenesis(context.api, null)) as GenesisCommit
+  it('Takes controller from authenticated DID if controller not specified', async () => {
+    const commit = (await ModelInstanceDocument.makeGenesis(context.api, null)) as GenesisCommit
     expect(commit.header.controllers[0]).toEqual(did.id)
   })
 
-  it('Takes controller from authenticated DID if controller not specified', async () => {
-    const signedCommitWithContent = await TileDocument.makeGenesis(
+  it('Does not sign commit if no content', async () => {
+    const signedCommitWithContent = await ModelInstanceDocument.makeGenesis(
       context.api,
       COMMITS.genesis.data
     )
@@ -353,7 +347,7 @@ describe('TileDocumentHandler', () => {
     expect(payload.data).toEqual(COMMITS.genesis.data)
     expect(payload.header.controllers[0]).toEqual(did.id)
 
-    const commitWithoutContent = (await TileDocument.makeGenesis(
+    const commitWithoutContent = (await ModelInstanceDocument.makeGenesis(
       context.api,
       null
     )) as GenesisCommit
@@ -362,33 +356,28 @@ describe('TileDocumentHandler', () => {
   })
 
   it('throws if more than one controller', async () => {
-    const commit1Promised = TileDocument.makeGenesis(context.api, COMMITS.genesis.data, {
+    const commit1Promised = ModelInstanceDocument.makeGenesis(context.api, COMMITS.genesis.data, {
       controllers: [did.id, 'did:key:zQ3shwsCgFanBax6UiaLu1oGvM7vhuqoW88VBUiUTCeHbTeTV'],
-      deterministic: true,
     })
     await expect(commit1Promised).rejects.toThrow(/Exactly one controller must be specified/)
   })
 
   it('creates genesis commits uniquely by default', async () => {
-    const commit1 = await TileDocument.makeGenesis(context.api, COMMITS.genesis.data)
-    const commit2 = await TileDocument.makeGenesis(context.api, COMMITS.genesis.data)
+    const commit1 = await ModelInstanceDocument.makeGenesis(context.api, COMMITS.genesis.data)
+    const commit2 = await ModelInstanceDocument.makeGenesis(context.api, COMMITS.genesis.data)
 
     expect(commit1).not.toEqual(commit2)
   })
 
-  it('creates genesis commits deterministically if deterministic:true is specified', async () => {
-    const metadata = { deterministic: true, controllers: ['a'], family: 'family', tags: ['x', 'y'] }
-    const commit1 = await TileDocument.makeGenesis(context.api, null, metadata)
-    const commit2 = await TileDocument.makeGenesis(context.api, null, metadata)
-
-    expect(commit1).toEqual(commit2)
-  })
-
   it('creates genesis commits without DID if content is undefined', async () => {
     await expect(
-      TileDocument.makeGenesis({} as CeramicApi, { foo: 'asdf' }, { controllers: [did.id] })
+      ModelInstanceDocument.makeGenesis(
+        {} as CeramicApi,
+        { foo: 'asdf' },
+        { controllers: [did.id] }
+      )
     ).rejects.toThrow('No DID provided')
-    const commit1 = await TileDocument.makeGenesis({} as CeramicApi, null, {
+    const commit1 = await ModelInstanceDocument.makeGenesis({} as CeramicApi, null, {
       controllers: [did.id],
     })
 
@@ -396,9 +385,9 @@ describe('TileDocumentHandler', () => {
   })
 
   it('applies genesis commit correctly', async () => {
-    const tileHandler = new TileDocumentHandler()
+    const modelInstanceDocumentHandler = new ModelInstanceDocumentHandler()
 
-    const commit = (await TileDocument.makeGenesis(
+    const commit = (await ModelInstanceDocument.makeGenesis(
       context.api,
       COMMITS.genesis.data
     )) as SignedCommitContainer
@@ -413,13 +402,13 @@ describe('TileDocumentHandler', () => {
       commit: payload,
       envelope: commit.jws,
     }
-    const streamState = await tileHandler.applyCommit(commitData, context)
+    const streamState = await modelInstanceDocumentHandler.applyCommit(commitData, context)
     delete streamState.metadata.unique
     expect(streamState).toMatchSnapshot()
   })
 
   it('makes signed commit correctly', async () => {
-    const tileDocumentHandler = new TileDocumentHandler()
+    const modelInstanceDocumentHandler = new ModelInstanceDocumentHandler()
 
     await context.ipfs.dag.put(COMMITS.genesisGenerated.jws, FAKE_CID_1)
 
@@ -434,9 +423,9 @@ describe('TileDocumentHandler', () => {
       commit: COMMITS.genesisGenerated.linkedBlock,
       envelope: COMMITS.genesisGenerated.jws,
     }
-    const state = await tileDocumentHandler.applyCommit(commitData, context)
+    const state = await modelInstanceDocumentHandler.applyCommit(commitData, context)
     const state$ = TestUtils.runningState(state)
-    const doc = new TileDocument(state$, context)
+    const doc = new ModelInstanceDocument(state$, context)
 
     await expect(doc.makeCommit({} as CeramicApi, COMMITS.r1.desiredContent)).rejects.toThrow(
       /No DID/
@@ -452,9 +441,9 @@ describe('TileDocumentHandler', () => {
   })
 
   it('applies signed commit correctly', async () => {
-    const tileDocumentHandler = new TileDocumentHandler()
+    const modelInstanceDocumentHandler = new ModelInstanceDocumentHandler()
 
-    const genesisCommit = (await TileDocument.makeGenesis(
+    const genesisCommit = (await ModelInstanceDocument.makeGenesis(
       context.api,
       COMMITS.genesis.data
     )) as SignedCommitContainer
@@ -470,10 +459,10 @@ describe('TileDocumentHandler', () => {
       commit: payload,
       envelope: genesisCommit.jws,
     }
-    let state = await tileDocumentHandler.applyCommit(genesisCommitData, context)
+    let state = await modelInstanceDocumentHandler.applyCommit(genesisCommitData, context)
 
     const state$ = TestUtils.runningState(state)
-    const doc = new TileDocument(state$, context)
+    const doc = new ModelInstanceDocument(state$, context)
     const signedCommit = (await doc.makeCommit(
       context.api,
       COMMITS.r1.desiredContent
@@ -491,7 +480,7 @@ describe('TileDocumentHandler', () => {
       commit: sPayload,
       envelope: signedCommit.jws,
     }
-    state = await tileDocumentHandler.applyCommit(signedCommitData, context, state)
+    state = await modelInstanceDocumentHandler.applyCommit(signedCommitData, context, state)
     delete state.metadata.unique
     delete state.next.metadata.unique
     expect(state).toMatchSnapshot()
@@ -499,9 +488,9 @@ describe('TileDocumentHandler', () => {
 
   it('multiple consecutive updates', async () => {
     const deepCopy = (o) => StreamUtils.deserializeState(StreamUtils.serializeState(o))
-    const tileDocumentHandler = new TileDocumentHandler()
+    const modelInstanceDocumentHandler = new ModelInstanceDocumentHandler()
 
-    const genesisCommit = (await TileDocument.makeGenesis(context.api, {
+    const genesisCommit = (await ModelInstanceDocument.makeGenesis(context.api, {
       test: 'data',
     })) as SignedCommitContainer
     await context.ipfs.dag.put(genesisCommit, FAKE_CID_1)
@@ -514,11 +503,11 @@ describe('TileDocumentHandler', () => {
       commit: payload,
       envelope: genesisCommit.jws,
     }
-    const genesisState = await tileDocumentHandler.applyCommit(genesisCommitData, context)
+    const genesisState = await modelInstanceDocumentHandler.applyCommit(genesisCommitData, context)
 
     // make a first update
     const state$ = TestUtils.runningState(genesisState)
-    let doc = new TileDocument(state$, context)
+    let doc = new ModelInstanceDocument(state$, context)
     const signedCommit1 = (await doc.makeCommit(context.api, {
       other: { obj: 'content' },
     })) as SignedCommitContainer
@@ -533,7 +522,7 @@ describe('TileDocumentHandler', () => {
       commit: sPayload1,
       envelope: signedCommit1.jws,
     }
-    const state1 = await tileDocumentHandler.applyCommit(
+    const state1 = await modelInstanceDocumentHandler.applyCommit(
       signedCommitData_1,
       context,
       deepCopy(genesisState)
@@ -541,7 +530,7 @@ describe('TileDocumentHandler', () => {
 
     // make a second update on top of the first
     const state1$ = TestUtils.runningState(state1)
-    doc = new TileDocument(state1$, context)
+    doc = new ModelInstanceDocument(state1$, context)
     const signedCommit2 = (await doc.makeCommit(context.api, {
       other: { obj2: 'fefe' },
     })) as SignedCommitContainer
@@ -557,7 +546,7 @@ describe('TileDocumentHandler', () => {
       commit: sPayload2,
       envelope: signedCommit2.jws,
     }
-    const state2 = await tileDocumentHandler.applyCommit(
+    const state2 = await modelInstanceDocumentHandler.applyCommit(
       signedCommitData_2,
       context,
       deepCopy(state1)
@@ -568,11 +557,15 @@ describe('TileDocumentHandler', () => {
   })
 
   it('throws error if commit signed by wrong DID', async () => {
-    const tileDocumentHandler = new TileDocumentHandler()
+    const modelInstanceDocumentHandler = new ModelInstanceDocumentHandler()
 
-    const genesisCommit = (await TileDocument.makeGenesis(context.api, COMMITS.genesis.data, {
-      controllers: ['did:3:fake'],
-    })) as SignedCommitContainer
+    const genesisCommit = (await ModelInstanceDocument.makeGenesis(
+      context.api,
+      COMMITS.genesis.data,
+      {
+        controllers: ['did:3:fake'],
+      }
+    )) as SignedCommitContainer
     await context.ipfs.dag.put(genesisCommit, FAKE_CID_1)
 
     const payload = dagCBOR.decode(genesisCommit.linkedBlock)
@@ -585,17 +578,21 @@ describe('TileDocumentHandler', () => {
       envelope: genesisCommit.jws,
       timestamp: Date.now(),
     }
-    await expect(tileDocumentHandler.applyCommit(genesisCommitData, context)).rejects.toThrow(
-      /invalid_jws: not a valid verificationMethod for issuer/
-    )
+    await expect(
+      modelInstanceDocumentHandler.applyCommit(genesisCommitData, context)
+    ).rejects.toThrow(/invalid_jws: not a valid verificationMethod for issuer/)
   })
 
   it('throws error if changes to more than one controller', async () => {
-    const tileDocumentHandler = new TileDocumentHandler()
+    const modelInstanceDocumentHandler = new ModelInstanceDocumentHandler()
 
-    const genesisCommit = (await TileDocument.makeGenesis(context.api, COMMITS.genesis.data, {
-      controllers: [did.id],
-    })) as SignedCommitContainer
+    const genesisCommit = (await ModelInstanceDocument.makeGenesis(
+      context.api,
+      COMMITS.genesis.data,
+      {
+        controllers: [did.id],
+      }
+    )) as SignedCommitContainer
     await context.ipfs.dag.put(genesisCommit, FAKE_CID_1)
 
     const payload = dagCBOR.decode(genesisCommit.linkedBlock)
@@ -607,8 +604,8 @@ describe('TileDocumentHandler', () => {
       commit: payload,
       envelope: genesisCommit.jws,
     }
-    const state = await tileDocumentHandler.applyCommit(genesisCommitData, context)
-    const doc = new TileDocument(state, context)
+    const state = await modelInstanceDocumentHandler.applyCommit(genesisCommitData, context)
+    const doc = new ModelInstanceDocument(state, context)
     const makeCommit = doc.makeCommit(context.api, COMMITS.r1.desiredContent, {
       controllers: [did.id, did.id],
     })
@@ -616,9 +613,9 @@ describe('TileDocumentHandler', () => {
   })
 
   it('applies anchor commit correctly', async () => {
-    const tileDocumentHandler = new TileDocumentHandler()
+    const modelInstanceDocumentHandler = new ModelInstanceDocumentHandler()
 
-    const genesisCommit = (await TileDocument.makeGenesis(
+    const genesisCommit = (await ModelInstanceDocument.makeGenesis(
       context.api,
       COMMITS.genesis.data
     )) as SignedCommitContainer
@@ -634,10 +631,10 @@ describe('TileDocumentHandler', () => {
       commit: payload,
       envelope: genesisCommit.jws,
     }
-    let state = await tileDocumentHandler.applyCommit(genesisCommitData, context)
+    let state = await modelInstanceDocumentHandler.applyCommit(genesisCommitData, context)
 
     const state$ = TestUtils.runningState(state)
-    const doc = new TileDocument(state$, context)
+    const doc = new ModelInstanceDocument(state$, context)
     const signedCommit = (await doc.makeCommit(
       context.api,
       COMMITS.r1.desiredContent
@@ -655,7 +652,7 @@ describe('TileDocumentHandler', () => {
       commit: sPayload,
       envelope: signedCommit.jws,
     }
-    state = await tileDocumentHandler.applyCommit(signedCommitData, context, state)
+    state = await modelInstanceDocumentHandler.applyCommit(signedCommitData, context, state)
 
     await context.ipfs.dag.put(COMMITS.proof, FAKE_CID_4)
     // apply anchor
@@ -665,7 +662,7 @@ describe('TileDocumentHandler', () => {
       commit: COMMITS.r2.commit,
       proof: COMMITS.proof,
     }
-    state = await tileDocumentHandler.applyCommit(anchorCommitData, context, state)
+    state = await modelInstanceDocumentHandler.applyCommit(anchorCommitData, context, state)
     delete state.metadata.unique
     expect(state).toMatchSnapshot()
   })
@@ -673,10 +670,10 @@ describe('TileDocumentHandler', () => {
   it('fails to apply commit if old key is used to make the commit and keys have been rotated', async () => {
     const rotateDate = new Date('2022-03-11T21:28:07.383Z')
 
-    const tileDocumentHandler = new TileDocumentHandler()
+    const modelInstanceDocumentHandler = new ModelInstanceDocumentHandler()
 
     // make and apply genesis with old key
-    const genesisCommit = (await TileDocument.makeGenesis(
+    const genesisCommit = (await ModelInstanceDocument.makeGenesis(
       signerUsingOldKey,
       COMMITS.genesis.data
     )) as SignedCommitContainer
@@ -694,13 +691,13 @@ describe('TileDocumentHandler', () => {
       timestamp: rotateDate.valueOf() / 1000 - 60 * 60,
     }
 
-    const state = await tileDocumentHandler.applyCommit(genesisCommitData, context)
+    const state = await modelInstanceDocumentHandler.applyCommit(genesisCommitData, context)
 
     rotateKey(did, rotateDate.toISOString())
 
     // make update with old key
     const state$ = TestUtils.runningState(state)
-    const doc = new TileDocument(state$, context)
+    const doc = new ModelInstanceDocument(state$, context)
     const signedCommit = (await doc.makeCommit(
       signerUsingOldKey,
       COMMITS.r1.desiredContent
@@ -721,18 +718,18 @@ describe('TileDocumentHandler', () => {
     }
 
     // applying a commit made with the old key after rotation
-    await expect(tileDocumentHandler.applyCommit(signedCommitData, context, state)).rejects.toThrow(
-      /invalid_jws: signature authored with a revoked DID version/
-    )
+    await expect(
+      modelInstanceDocumentHandler.applyCommit(signedCommitData, context, state)
+    ).rejects.toThrow(/invalid_jws: signature authored with a revoked DID version/)
   })
 
   it('fails to apply commit if new key used before rotation', async () => {
     const rotateDate = new Date('2022-03-11T21:28:07.383Z')
 
-    const tileDocumentHandler = new TileDocumentHandler()
+    const modelInstanceDocumentHandler = new ModelInstanceDocumentHandler()
 
     // make genesis with new key
-    const genesisCommit = (await TileDocument.makeGenesis(
+    const genesisCommit = (await ModelInstanceDocument.makeGenesis(
       signerUsingNewKey,
       COMMITS.genesis.data
     )) as SignedCommitContainer
@@ -752,19 +749,19 @@ describe('TileDocumentHandler', () => {
 
     rotateKey(did, rotateDate.toISOString())
 
-    await expect(tileDocumentHandler.applyCommit(genesisCommitData, context)).rejects.toThrow(
-      /invalid_jws: signature authored before creation of DID version/
-    )
+    await expect(
+      modelInstanceDocumentHandler.applyCommit(genesisCommitData, context)
+    ).rejects.toThrow(/invalid_jws: signature authored before creation of DID version/)
   })
 
   it('applies commit made using an old key if it is applied within the revocation period', async () => {
     const rotateDate = new Date('2022-03-11T21:28:07.383Z')
     rotateKey(did, rotateDate.toISOString())
 
-    const tileDocumentHandler = new TileDocumentHandler()
+    const modelInstanceDocumentHandler = new ModelInstanceDocumentHandler()
 
     // make genesis commit using old key
-    const genesisCommit = (await TileDocument.makeGenesis(
+    const genesisCommit = (await ModelInstanceDocument.makeGenesis(
       signerUsingOldKey,
       COMMITS.genesis.data
     )) as SignedCommitContainer
@@ -781,22 +778,9 @@ describe('TileDocumentHandler', () => {
       envelope: genesisCommit.jws,
       timestamp: rotateDate.valueOf() / 1000 + 60 * 60,
     }
-    const state = await tileDocumentHandler.applyCommit(genesisCommitData, context)
+    const state = await modelInstanceDocumentHandler.applyCommit(genesisCommitData, context)
     delete state.metadata.unique
 
     expect(state).toMatchSnapshot()
-  })
-})
-
-describe('TileHandler', () => {
-  test('can not create invalid deterministic tile document', async () => {
-    const fauxCeramic = {} as unknown as CeramicApi
-    await expect(
-      TileDocument.makeGenesis(fauxCeramic, undefined, {
-        controllers: ['did:foo:blah'],
-        family: 'test123',
-        tags: ['foo', undefined, 'blah'],
-      })
-    ).rejects.toThrow(/`undefined` is not supported by the IPLD Data Model and cannot be encoded/)
   })
 })