+++ packages/stream-model-instance-handler/src/model-instance-document-handler.ts
@@ -1,0 +1,247 @@
+import jsonpatch from 'fast-json-patch'
+import cloneDeep from 'lodash.clonedeep'
+import { ModelInstanceDocument } from '@ceramicnetwork/stream-model-instance'
+import {
+  AnchorStatus,
+  CommitData,
+  CommitType,
+  Context,
+  SignatureStatus,
+  SignatureUtils,
+  StreamConstructor,
+  StreamHandler,
+  StreamState,
+  StreamUtils,
+} from '@ceramicnetwork/common'
+import { StreamID } from '@ceramicnetwork/streamid'
+import { SchemaValidation } from './schema-utils.js'
+
+function stringArraysEqual(arr1: Array<string>, arr2: Array<string>) {
+  if (arr1.length != arr2.length) {
+    return false
+  }
+  for (let i = 0; i < arr1.length; i++) {
+    if (arr1[i] !== arr2[i]) {
+      return false
+    }
+  }
+  return true
+}
+
+/**
+ * ModelInstanceDocument stream handler implementation
+ */
+export class ModelInstanceDocumentHandler implements StreamHandler<ModelInstanceDocument> {
+  private readonly _schemaValidator: SchemaValidation
+
+  constructor() {
+    this._schemaValidator = new SchemaValidation()
+  }
+
+  get type(): number {
+    return ModelInstanceDocument.STREAM_TYPE_ID
+  }
+
+  get name(): string {
+    return ModelInstanceDocument.STREAM_TYPE_NAME
+  }
+
+  get stream_constructor(): StreamConstructor<ModelInstanceDocument> {
+    return ModelInstanceDocument
+  }
+
+  /**
+   * Applies commit (genesis|signed|anchor)
+   * @param commitData - Commit (with JWS envelope or anchor proof, if available and extracted before application)
+   * @param context - Ceramic context
+   * @param state - Document state
+   */
+  async applyCommit(
+    commitData: CommitData,
+    context: Context,
+    state?: StreamState
+  ): Promise<StreamState> {
+    if (state == null) {
+      // apply genesis
+      return this._applyGenesis(commitData, context)
+    }
+
+    if (StreamUtils.isAnchorCommitData(commitData)) {
+      return this._applyAnchor(context, commitData, state)
+    }
+
+    return this._applySigned(commitData, state, context)
+  }
+
+  /**
+   * Applies genesis commit
+   * @param commitData - Genesis commit
+   * @param context - Ceramic context
+   * @private
+   */
+  async _applyGenesis(commitData: CommitData, context: Context): Promise<StreamState> {
+    const payload = commitData.commit
+    const isSigned = StreamUtils.isSignedCommitData(commitData)
+    if (isSigned) {
+      const streamId = await StreamID.fromGenesis('MID', commitData.commit)
+      const { controllers } = payload.header
+      // TODO(NET-1437): replace family with model
+      await SignatureUtils.verifyCommitSignature(
+        commitData,
+        context.did,
+        controllers[0],
+        null,
+        streamId
+      )
+    } else if (payload.data) {
+      throw Error('Genesis commit with contents should always be signed')
+    }
+
+    if (!(payload.header.controllers && payload.header.controllers.length === 1)) {
+      throw new Error('Exactly one controller must be specified')
+    }
+
+    const state = {
+      type: ModelInstanceDocument.STREAM_TYPE_ID,
+      content: payload.data || {},
+      metadata: payload.header,
+      signature: isSigned ? SignatureStatus.SIGNED : SignatureStatus.GENESIS,
+      anchorStatus: AnchorStatus.NOT_REQUESTED,
+      log: [{ cid: commitData.cid, type: CommitType.GENESIS }],
+    }
+
+    // TODO: re-enable once model schema validation was added
+    /*if (state.metadata.schema) {
+      await this._schemaValidator.validateSchema(context.api, state.content, state.metadata.schema)
+    }*/
+
+    return state
+  }
+
+  /**
+   * Applies signed commit
+   * @param commitData - Signed commit
+   * @param state - Document state
+   * @param context - Ceramic context
+   * @private
+   */
+  async _applySigned(
+    commitData: CommitData,
+    state: StreamState,
+    context: Context
+  ): Promise<StreamState> {
+    const controller = state.next?.metadata?.controllers?.[0] || state.metadata.controllers[0]
+
+    // Verify the signature first
+    const streamId = StreamUtils.streamIdFromState(state)
+    // TODO(NET-1437): replace family with model
+    await SignatureUtils.verifyCommitSignature(
+      commitData,
+      context.did,
+      controller,
+      null,
+      streamId
+    )
+
+    // Retrieve the payload
+    const payload = commitData.commit
+
+    if (!payload.id.equals(state.log[0].cid)) {
+      throw new Error(`Invalid streamId ${payload.id}, expected ${state.log[0].cid}`)
+    }
+    const expectedPrev = state.log[state.log.length - 1].cid
+    if (!payload.prev.equals(expectedPrev)) {
+      throw new Error(
+        `Commit doesn't properly point to previous commit in log. Expected ${expectedPrev}, found 'prev' ${payload.prev}`
+      )
+    }
+
+    if (payload.header.controllers && payload.header.controllers.length !== 1) {
+      throw new Error('Exactly one controller must be specified')
+    }
+
+    if (
+      state.metadata.forbidControllerChange &&
+      payload.header.controllers &&
+      !stringArraysEqual(payload.header.controllers, state.metadata.controllers)
+    ) {
+      const streamId = new StreamID(ModelInstanceDocument.STREAM_TYPE_ID, state.log[0].cid)
+      throw new Error(
+        `Cannot change controllers since 'forbidControllerChange' is set. Tried to change controllers for Stream ${streamId} from ${JSON.stringify(
+          state.metadata.controllers
+        )} to ${payload.header.controllers}`
+      )
+    }
+
+    if (
+      payload.header.forbidControllerChange !== undefined &&
+      payload.header.forbidControllerChange !== state.metadata.forbidControllerChange
+    ) {
+      throw new Error("Changing 'forbidControllerChange' metadata property is not allowed")
+    }
+
+    const nextState = cloneDeep(state)
+
+    nextState.signature = SignatureStatus.SIGNED
+    nextState.anchorStatus = AnchorStatus.NOT_REQUESTED
+
+    nextState.log.push({ cid: commitData.cid, type: CommitType.SIGNED })
+
+    const oldContent = state.next?.content ?? state.content
+    const oldMetadata = state.next?.metadata ?? state.metadata
+
+    const newContent = jsonpatch.applyPatch(oldContent, payload.data).newDocument
+    const newMetadata = { ...oldMetadata, ...payload.header }
+
+    nextState.next = {
+      content: newContent,
+      metadata: newMetadata,
+    }
+
+    return nextState
+  }
+
+  /**
+   * Applies anchor commit
+   * @param context - Ceramic context
+   * @param commitData - Anchor commit
+   * @param state - Document state
+   * @private
+   */
+  async _applyAnchor(
+    context: Context,
+    commitData: CommitData,
+    state: StreamState
+  ): Promise<StreamState> {
+    // TODO: Assert that the 'prev' of the commit being applied is the end of the log in 'state'
+    const proof = commitData.proof
+    state.log.push({
+      cid: commitData.cid,
+      type: CommitType.ANCHOR,
+      timestamp: proof.blockTimestamp,
+    })
+    let content = state.content
+    let metadata = state.metadata
+
+    if (state.next?.content) {
+      content = state.next.content
+      delete state.next.content
+    }
+
+    if (state.next?.metadata) {
+      metadata = state.next.metadata
+      delete state.next.metadata
+    }
+
+    delete state.next
+    delete state.anchorScheduledFor
+
+    return {
+      ...state,
+      content,
+      metadata,
+      anchorStatus: AnchorStatus.ANCHORED,
+      anchorProof: proof,
+    }
+  }
+}