+++ packages/stream-tile/src/tile-document.ts
@@ -138,20 +138,6 @@ async function _ensureAuthenticated(signer: CeramicSigner) {
   }
 }
 
-/**
- * Sign a Tile commit with the currently authenticated DID.
- * @param signer - Object containing the DID to use to sign the commit
- * @param commit - Commit to be signed
- * @private
- */
-async function _signDagJWS(
-  signer: CeramicSigner,
-  commit: CeramicCommit
-): Promise<SignedCommitContainer> {
-  await _ensureAuthenticated(signer)
-  return signer.did.createDagJWS(commit)
-}
-
 async function throwReadOnlyError(): Promise<void> {
   throw new Error(
     'Historical stream commits cannot be modified. Load the stream without specifying a commit to make updates.'
@@ -217,7 +203,9 @@ export class TileDocument<T = Record<string, any>> extends Stream {
       // document as there shouldn't be any existing state for this doc on the network.
       opts.syncTimeoutSeconds = 0
     }
-    const commit = genesisCommit.data ? await _signDagJWS(ceramic, genesisCommit) : genesisCommit
+    const commit = genesisCommit.data
+      ? await TileDocument._signDagJWS(ceramic, genesisCommit)
+      : genesisCommit
     return ceramic.createStreamFromGenesis<TileDocument<T>>(
       TileDocument.STREAM_TYPE_ID,
       commit,
@@ -307,7 +295,7 @@ export class TileDocument<T = Record<string, any>> extends Stream {
       prev: this.tip,
       id: this.state$.id.cid,
     }
-    const commit = await _signDagJWS(this.api, rawCommit)
+    const commit = await TileDocument._signDagJWS(this.api, rawCommit)
     const updated = await this.api.applyCommit(this.id, commit, opts)
     this.state$.next(updated.state)
   }
@@ -338,14 +326,35 @@ export class TileDocument<T = Record<string, any>> extends Stream {
     newContent: T | null | undefined,
     newMetadata?: TileMetadataArgs
   ): Promise<CeramicCommit> {
+    const commit = await this._makeRawCommit(signer, newContent, newMetadata)
+    return TileDocument._signDagJWS(signer, commit)
+  }
+
+  /**
+   * Helper function for makeCommit() to sanity check input values and
+   * allow unit tests to update newMetadata before signing.
+   * @param signer - Object containing the DID making (and signing) the commit
+   * @param newContent
+   * @param newMetadata
+   */
+  private async _makeRawCommit(
+    signer: CeramicSigner,
+    newContent: T | null | undefined,
+    newMetadata?: TileMetadataArgs
+  ): Promise<RawCommit> {
     const header = headerFromMetadata(newMetadata, false)
 
     if (newContent == null) {
       newContent = this.content
     }
 
-    if (header.controllers && header.controllers?.length !== 1) {
-      throw new Error('Exactly one controller must be specified')
+    if (header.controllers) {
+      if (header.controllers.length !== 1) {
+        throw new Error('Exactly one controller must be specified')
+      }
+      if (!header.controllers[0]) {
+        throw new Error('Controller cannot be updated to an undefined value.')
+      }
     }
 
     const patch = jsonpatch.compare(this.content, newContent)
@@ -355,7 +364,7 @@ export class TileDocument<T = Record<string, any>> extends Stream {
       prev: this.tip,
       id: this.state.log[0].cid,
     }
-    return _signDagJWS(signer, commit)
+    return commit
   }
 
   /**
@@ -376,7 +385,7 @@ export class TileDocument<T = Record<string, any>> extends Stream {
     if (!metadata.controllers || metadata.controllers.length === 0) {
       if (signer.did) {
         await _ensureAuthenticated(signer)
-        // When did has parent, it has a capability, the did issuer (parent) of the capability is the stream controller 
+        // When did has parent, it has a capability, the did issuer (parent) of the capability is the stream controller
         metadata.controllers = [signer.did.hasParent ? signer.did.parent : signer.did.id]
       } else {
         throw new Error('No controllers specified')
@@ -399,6 +408,20 @@ export class TileDocument<T = Record<string, any>> extends Stream {
       return result
     }
     const commit: GenesisCommit = { data: content, header }
-    return _signDagJWS(signer, commit)
+    return TileDocument._signDagJWS(signer, commit)
+  }
+
+  /**
+   * Sign a Tile commit with the currently authenticated DID.
+   * @param signer - Object containing the DID to use to sign the commit
+   * @param commit - Commit to be signed
+   * @private
+   */
+  private static async _signDagJWS(
+    signer: CeramicSigner,
+    commit: CeramicCommit
+  ): Promise<SignedCommitContainer> {
+    await _ensureAuthenticated(signer)
+    return signer.did.createDagJWS(commit)
   }
 }